requirements.txt
asgiref==3.10.0
beautifulsoup4==4.14.2
bootstrap4==0.1.0
boto3==1.40.70
botocore==1.40.70
dj-database-url==3.0.1
Django==5.2.8
django-admin-sortable2==2.2.8
django-bootstrap4==25.2
django-ckeditor==6.7.3
django-extensions==4.1
django-js-asset==3.1.2
django-storages==1.14.6
django-tinymce==5.0.0
dotenv==0.9.9
gunicorn==23.0.0
jmespath==1.0.1
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
s3transfer==0.14.0
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
whitenoise==6.11.0
